refactor(navigator): use navigator poll to load mission cache and move full mission cache into its own class
