+++ src/modules/navigator/Kconfig
@@ -71,5 +71,6 @@ config NAVIGATOR_FULL_MISSION_CACHE_SIZE
 		Maximum number of mission items reserved in RAM for full-route planning helpers:
 		56 bytes per item, allocated once at Navigator start. Set to 0 to disable the
 		cache on boards without the extra RAM. Longer missions are not cached at all.
-		After upload or reboot the cache refills in the background: about one item per
-		Navigator cycle (~20/s) while armed, or via short blocking reads while disarmed.
+		After upload or reboot the cache refills asynchronously. Navigator polls the
+		Dataman response subscription while each read is pending, then immediately
+		queues the next item without blocking its main loop.
