+++ src/modules/navigator/mission_route_cache.cpp
@@ -33,8 +33,8 @@
 /**
  * @file mission_route_cache.cpp
  *
- * Dataman-backed Provider for mission-route planning. It preloads the whole
- * uploaded mission, the published mission land item, and the safe points so
+ * Dataman-backed Provider for mission-route planning. It composes the optional
+ * full-mission cache with the published mission land item and safe-point caches so
  * the planner can scan route geometry without blocking Navigator on dataman or
  * SD-card reads. Normal mission execution can keep using smaller sliding caches;
  * use this cache when a caller needs access to the full route.
@@ -46,9 +46,7 @@
 
 #include "mission_route_types.h"
 
-#include <px4_platform_common/events.h>
 #include <px4_platform_common/log.h>
-#include <systemlib/mavlink_log.h>
 
 // The mission and land caches only accept these dataman storages as their source.
 static bool isMissionDatamanIdValid(uint8_t dataman_id)
@@ -57,109 +55,28 @@ static bool isMissionDatamanIdValid(uint8_t dataman_id)
 }
 
 MissionRouteCache::MissionRouteCache(orb_advert_t *mavlink_log_pub)
+	: _full_mission_cache(mavlink_log_pub)
 {
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-	_mavlink_log_pub = mavlink_log_pub;
-
-	// Allocated once at Navigator start and kept for the module lifetime. Intentionally
-	// not zero-filled: no index is exposed before a successful dataman read fills it.
-	_mission_items = new mission_item_s[kMaxFullMissionCacheSize];
-
-	if (_mission_items == nullptr) {
-		PX4_ERR("Mission cache allocation failed");
-	}
-
-#else
-	(void)mavlink_log_pub;
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
 }
 
-MissionRouteCache::~MissionRouteCache()
-{
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-	_dataman_client_mission.abortCurrentOperation();
-	delete[] _mission_items;
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
-}
+MissionRouteCache::~MissionRouteCache() = default;
 
-void MissionRouteCache::update(const mission_s &mission, bool allow_blocking_load)
+void MissionRouteCache::update(const mission_s &mission)
 {
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-	updateMissionCache(mission, allow_blocking_load);
-#else
-	(void)allow_blocking_load;
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
+	_full_mission_cache.update(mission);
 	updateMissionLandItemCache(mission);
 	updateSafePointCache(mission);
 }
 
 void MissionRouteCache::invalidate()
 {
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-	advanceMissionGeneration();
-	_mission = {};
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
+	_full_mission_cache.invalidate();
 
 	_mission_land = {};
 	_dataman_cache_land_item.invalidate();
 
 	resetSafePointCacheState(true);
 }
-
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-void MissionRouteCache::advanceMissionGeneration()
-{
-	++_mission_generation;
-
-	// Zero identifies a view that has never been published.
-	if (_mission_generation == 0) {
-		++_mission_generation;
-	}
-}
-
-// Start the load for a new source, which must be valid and loadable.
-// Read failures do not restart here, they resume at the failed index.
-void MissionRouteCache::startMissionLoad()
-{
-	// A rebuild immediately hides the old generation. An active older request
-	// is allowed to finish, and its result is discarded before this generation
-	// starts writing the shared typed buffer.
-	_mission.ready = false;
-	_mission.next_index = 0;
-	_mission.validation_pending = true;
-	_mission.retry.retry_at = 0;
-}
-
-bool MissionRouteCache::blockingLoadMissionItems(MissionCacheState &state)
-{
-	// Disarmed-only warm-up: synchronous reads with a bounded budget per update() call.
-	// readSync() polls in fixed 100 ms chunks, so a slow read can overrun kBlockingLoadItemTimeout.
-	const hrt_abstime start = hrt_absolute_time();
-
-	while (state.next_index < state.count) {
-		if (hrt_elapsed_time(&start) >= kBlockingLoadBudget) {
-			return false; // Out of budget, resume on the next update() call.
-		}
-
-		const int32_t request_index = state.next_index;
-
-		if (_dataman_client_mission.readSync(static_cast<dm_item_t>(state.dataman_id),
-						     static_cast<uint32_t>(request_index),
-						     reinterpret_cast<uint8_t *>(&_mission_items[request_index]),
-						     sizeof(mission_item_s), kBlockingLoadItemTimeout)) {
-			++state.next_index;
-
-		} else {
-			state.validation_pending = false;
-			PX4_WARN("Mission cache retry");
-			state.retry.scheduleRetry(hrt_absolute_time());
-			return false;
-		}
-	}
-
-	return true;
-}
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
 
 bool MissionRouteCache::missionLandItemCacheFullyLoaded() const
 {
@@ -235,61 +152,30 @@ void MissionRouteCache::resetSafePointCacheState(bool clear_source_identity)
 	_safe_point.source_dataman_id = source_dataman_id;
 }
 
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-// True while the current mission generation is fully loaded and readable.
-bool MissionRouteCache::missionCacheAvailable() const
-{
-	return _mission.initialized && _mission.source_valid && !_mission.too_large && _mission.ready;
-}
-
 bool MissionRouteCache::missionItemsReady(const mission_s &mission) const
 {
-	return missionCacheAvailable() && missionMatchesCache(mission);
+	return _full_mission_cache.missionItemsReady(mission);
 }
 
 int MissionRouteCache::missionCount() const
 {
-	return missionCacheAvailable() ? _mission.count : 0;
+	return _full_mission_cache.missionCount();
 }
 
 bool MissionRouteCache::loadMissionItem(int index, mission_item_s &mission_item) const
 {
-	if (missionCacheAvailable() && index >= 0 && index < _mission.count) {
-		mission_item = _mission_items[index];
-		return true;
-	}
-
-	return false;
+	return _full_mission_cache.loadMissionItem(index, mission_item);
 }
 
 bool MissionRouteCache::getMissionView(const mission_s &mission, MissionView &view) const
 {
-	if (!missionItemsReady(mission)) {
-		return false;
-	}
-
-	view.items = (_mission.count > 0) ? _mission_items : nullptr;
-	view.count = _mission.count;
-	view.mission_id = _mission.id;
-	view.dataman_id = _mission.dataman_id;
-	view.generation = _mission_generation;
-	return true;
+	return _full_mission_cache.getMissionView(mission, view);
 }
 
 bool MissionRouteCache::missionViewStillValid(const MissionView &view) const
 {
-	// Generation equality pins source and content: every source change, invalidate(),
-	// reload, and patch advances it, and it is only observable while ready.
-	return view.generation != 0 && view.generation == _mission_generation && _mission.ready;
+	return _full_mission_cache.missionViewStillValid(view);
 }
-
-#else
-bool MissionRouteCache::missionItemsReady(const mission_s &) const { return false; }
-int MissionRouteCache::missionCount() const { return 0; }
-bool MissionRouteCache::loadMissionItem(int, mission_item_s &) const { return false; }
-bool MissionRouteCache::getMissionView(const mission_s &, MissionView &) const { return false; }
-bool MissionRouteCache::missionViewStillValid(const MissionView &) const { return false; }
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
 
 int MissionRouteCache::safePointCount() const
 {
@@ -339,39 +225,7 @@ MissionRouteCache::SyncResult MissionRouteCache::syncMissionItem(const mission_s
 	// The land-item cache exists on every board and validates against the published
 	// mission, so keep it coherent independently of the full-mission cache outcome.
 	syncMissionLandItem(mission, index, mission_item);
-
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-
-	if (!_mission.source_valid || _mission.too_large
-	    || !missionMatchesCache(mission) || index < 0 || index >= _mission.count) {
-		return SyncResult::kRejected;
-	}
-
-	if (_mission.ready) {
-		_mission_items[index] = mission_item;
-		advanceMissionGeneration();
-		return SyncResult::kPatched;
-	}
-
-	if (index < _mission.next_index) {
-		// Indices below the load front are stable, patch in place. Deferring would
-		// lose the write: a resumed load does not re-read the prefix.
-		_mission_items[index] = mission_item;
-		return SyncResult::kPatched;
-	}
-
-	if (_mission_request.pending && _mission_request.index == index) {
-		// This read sampled the index before the write, so its response would store the
-		// old item. Advance the generation to discard it and read the index again.
-		advanceMissionGeneration();
-	}
-
-	// Any other index is still unread and dataman already holds the new item, so the
-	// pending load picks it up on its own.
-	return SyncResult::kDeferred;
-#else
-	return SyncResult::kRejected;
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
+	return _full_mission_cache.syncMissionItem(mission, index, mission_item);
 }
 
 void MissionRouteCache::syncMissionLandItem(const mission_s &mission, int32_t index, const mission_item_s &mission_item)
@@ -413,133 +267,6 @@ void MissionRouteCache::syncMissionLandItem(const mission_s &mission, int32_t in
 		}
 	}
 }
-
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-void MissionRouteCache::updateMissionCache(const mission_s &mission, bool allow_blocking_load)
-{
-	MissionCacheState &state = _mission;
-
-	if (!missionMatchesCache(mission)) {
-		// A new source immediately invalidates the published view. Do not abort
-		// an older request: it must be consumed and discarded before the same
-		// Dataman key can safely be requested for this generation.
-		advanceMissionGeneration();
-		state = {};
-		state.id = mission.mission_id;
-		state.count = mission.count;
-		state.dataman_id = mission.mission_dataman_id;
-		state.initialized = true;
-		state.source_valid = mission.count == 0 || isMissionDatamanIdValid(mission.mission_dataman_id);
-		state.ready = state.source_valid && mission.count == 0;
-		state.too_large = mission.count > kMaxFullMissionCacheSize;
-
-		if (!state.source_valid) {
-			PX4_ERR("Mission cache: invalid dataman id");
-		}
-
-		if (state.source_valid && state.too_large) {
-			// Runs once per source change; the route cache stays unavailable for this mission.
-			if (_mavlink_log_pub != nullptr) {
-				mavlink_log_warning(_mavlink_log_pub, "Mission with %u items exceeds route cache capacity %d\t",
-						    mission.count, static_cast<int>(kMaxFullMissionCacheSize));
-			}
-
-			events::send<uint16_t, int32_t>(events::ID("navigator_route_cache_too_large"),
-			{events::Log::Warning, events::LogInternal::Info},
-			"Mission with {1} items exceeds the route cache capacity {2}",
-			mission.count, kMaxFullMissionCacheSize);
-		}
-
-		// The boot-time allocation failed, no recovery is possible for a non-empty mission.
-		if (state.source_valid && !state.too_large && mission.count > 0 && _mission_items == nullptr) {
-			PX4_ERR("Mission cache unavailable: boot-time allocation failed");
-			state.too_large = true;
-		}
-
-		if (state.source_valid && !state.too_large && mission.count > 0) {
-			startMissionLoad();
-		}
-	}
-
-	_dataman_client_mission.update();
-
-	if (_mission_request.pending) {
-		bool request_success = false;
-
-		if (!_dataman_client_mission.lastOperationCompleted(request_success)) {
-			return;
-		}
-
-		const bool current_generation = _mission_request.generation == _mission_generation;
-		const bool current_index = _mission_request.index == state.next_index;
-		_mission_request.pending = false;
-
-		if (current_generation && current_index && state.validation_pending) {
-			if (request_success) {
-				++state.next_index;
-
-			} else {
-				state.validation_pending = false;
-				PX4_WARN("Mission cache retry");
-				state.retry.scheduleRetry(hrt_absolute_time());
-				return;
-			}
-		}
-	}
-
-	if (!state.source_valid || state.ready || state.too_large || mission.count <= 0) {
-		return;
-	}
-
-	const hrt_abstime now = hrt_absolute_time();
-
-	if (state.validation_pending) {
-		if (_mission_request.pending) {
-			// The blocking load below uses readSync(), which would consume this pending
-			// response and stall the async read until its 100 ms self-retry.
-			return;
-		}
-
-		if (allow_blocking_load && state.next_index < state.count && !blockingLoadMissionItems(state)) {
-			// Budget exhausted (resume next call) or a failed read scheduled a retry.
-			return;
-		}
-
-		if (state.next_index < state.count) {
-			const int32_t request_index = state.next_index;
-			const bool request_started = _dataman_client_mission.readAsync(
-							     static_cast<dm_item_t>(state.dataman_id),
-							     static_cast<uint32_t>(request_index),
-							     reinterpret_cast<uint8_t *>(&_mission_items[request_index]),
-							     sizeof(mission_item_s));
-
-			if (request_started) {
-				_mission_request.generation = _mission_generation;
-				_mission_request.index = request_index;
-				_mission_request.pending = true;
-
-			} else {
-				state.validation_pending = false;
-				PX4_WARN("Mission cache read start failed");
-				state.retry.scheduleRetry(now);
-			}
-
-			return;
-		}
-
-		// Every failure path clears validation_pending, so reaching the end of the
-		// sequential load means the complete generation was read successfully.
-		state.validation_pending = false;
-		state.ready = true;
-		state.retry.clear();
-
-	} else if (state.retry.due(now)) {
-		// Resume at the failed index, the loaded prefix is still valid.
-		state.validation_pending = true;
-		state.retry.retry_at = 0;
-	}
-}
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
 
 void MissionRouteCache::updateMissionLandItemCache(const mission_s &mission)
 {
@@ -765,16 +492,6 @@ void MissionRouteCache::updateSafePointCache(const mission_s &mission)
 		}
 	}
 }
-
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-bool MissionRouteCache::missionMatchesCache(const mission_s &mission) const
-{
-	return _mission.initialized
-	       && mission.mission_id == _mission.id
-	       && mission.count == _mission.count
-	       && mission.mission_dataman_id == _mission.dataman_id;
-}
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
 
 bool MissionRouteCache::missionLandMatchesCache(const mission_s &mission) const
 {
