+++ src/modules/navigator/mission_route_cache.h
@@ -44,6 +44,7 @@
 #include <dataman_client/DatamanClient.hpp>
 #include <uORB/topics/mission.h>
 
+#include "full_mission_cache.h"
 #include "navigation.h"
 #include "mission_route_provider.h"
 
@@ -54,27 +55,12 @@ public:
 	static constexpr hrt_abstime kCacheOnlyLoadWait{0};
 	static constexpr hrt_abstime kCacheRetryBackoff{500000}; // 500 ms
 	static constexpr uint8_t kMaxRetryBackoffShift{3}; // Retry 3+: 500ms << 3 = 4000 ms.
-	// Disarmed-only warm-up: bounded synchronous dataman reads per update() call.
-	static constexpr hrt_abstime kBlockingLoadBudget{20000}; // 20 ms per update() call
-	static constexpr hrt_abstime kBlockingLoadItemTimeout{50000}; // 50 ms per item read
-	static constexpr int32_t kMaxFullMissionCacheSize{CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE};
+	static constexpr int32_t kMaxFullMissionCacheSize{FullMissionCache::kMaxMissionCacheSize};
 	static constexpr uint32_t kInitialSafePointCacheSize{0};
 	static constexpr uint32_t kInitialLandItemCacheSize{1};
 
-	struct MissionView {
-		const mission_item_s *items{nullptr};
-		int32_t count{0};
-		uint32_t mission_id{0};
-		uint8_t dataman_id{DM_KEY_WAYPOINTS_OFFBOARD_0};
-		uint32_t generation{0};
-	};
-
-	/** @brief Outcome of syncMissionItem() for the full-mission cache. */
-	enum class SyncResult : uint8_t {
-		kRejected,  ///< The mission cache did not apply the write: source mismatch, index out of range, or cache compiled out.
-		kPatched,   ///< The item was applied in place; any previously borrowed view is stale.
-		kDeferred   ///< The item is not loaded yet; the pending load reads it from dataman.
-	};
+	using MissionView = FullMissionCache::MissionView;
+	using SyncResult = FullMissionCache::SyncResult;
 
 	// The optional mavlink log advert (Navigator's) reports a mission exceeding the cache.
 	explicit MissionRouteCache(orb_advert_t *mavlink_log_pub = nullptr);
@@ -82,14 +68,15 @@ public:
 	MissionRouteCache(const MissionRouteCache &) = delete;
 	MissionRouteCache &operator=(const MissionRouteCache &) = delete;
 
-	/**
-	 * @brief Advance the cache state machines for the published mission.
-	 *
-	 * allow_blocking_load permits short synchronous dataman reads (kBlockingLoadBudget per
-	 * call) for the pending mission load. Only pass true while disarmed.
-	 */
-	void update(const mission_s &mission, bool allow_blocking_load = false);
+	/** @brief Advance the cache state machines for the published mission. */
+	void update(const mission_s &mission);
 	void invalidate();
+
+	/** Full-mission Dataman response subscription while an async read is pending. */
+	orb_sub_t fullMissionResponseSubscription() const
+	{
+		return _full_mission_cache.responseSubscription();
+	}
 
 	bool missionItemsReady(const mission_s &mission) const;
 	bool safePointsReady() const
@@ -180,7 +167,7 @@ private:
 		kError
 	};
 
-	// Shared retry/backoff bookkeeping for the mission, land, and safe-point caches.
+	// Shared retry/backoff bookkeeping for the land and safe-point caches.
 	struct RetryBackoff {
 		hrt_abstime retry_at{0};
 		uint8_t retry_count{0};
@@ -202,27 +189,6 @@ private:
 			}
 		}
 	};
-
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-	struct MissionCacheState {
-		uint32_t id{0};
-		int32_t count{0};
-		uint8_t dataman_id{DM_KEY_WAYPOINTS_OFFBOARD_0};
-		int32_t next_index{0}; ///< Sequential load front; items below it are loaded.
-		bool initialized{false};
-		bool source_valid{false};
-		bool ready{false};
-		bool too_large{false};
-		bool validation_pending{false};
-		RetryBackoff retry{};
-	};
-
-	struct MissionRequest {
-		uint32_t generation{0}; ///< Generation the read was issued for; a mismatch discards the response.
-		int32_t index{-1};      ///< Mission item index being read.
-		bool pending{false};    ///< generation and index are only meaningful while true.
-	};
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
 
 	struct MissionLandState {
 		uint32_t mission_id{0};
@@ -246,14 +212,6 @@ private:
 		RetryBackoff retry{};
 	};
 
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-	void updateMissionCache(const mission_s &mission, bool allow_blocking_load);
-	bool missionMatchesCache(const mission_s &mission) const;
-	bool missionCacheAvailable() const;
-	void startMissionLoad();
-	bool blockingLoadMissionItems(MissionCacheState &state);
-	void advanceMissionGeneration();
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
 	void updateMissionLandItemCache(const mission_s &mission);
 	void syncMissionLandItem(const mission_s &mission, int32_t index, const mission_item_s &mission_item);
 	bool missionLandMatchesCache(const mission_s &mission) const;
@@ -267,21 +225,11 @@ private:
 
 	// Navigator accesses MissionRouteCache from one serialized task loop. The
 	// typed mission view can therefore be borrowed for one synchronous planning call.
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-	mission_item_s *_mission_items {nullptr};
-	DatamanClient _dataman_client_mission{};
-	orb_advert_t *_mavlink_log_pub{nullptr};
-	// Keep outside of MissionCacheState: it must survive `_mission = {}` reset
-	uint32_t _mission_generation{0};
-	MissionRequest _mission_request{};
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
+	FullMissionCache _full_mission_cache;
 	mutable DatamanCache _dataman_cache_safepoint {"navigator_dm_cache_route_safepoint", kInitialSafePointCacheSize};
 	mutable DatamanCache _dataman_cache_land_item{"navigator_dm_cache_route_land", kInitialLandItemCacheSize};
 	DatamanClient &_dataman_client_safepoint = _dataman_cache_safepoint.client();
 
-#if CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE > 0
-	MissionCacheState _mission {};
-#endif // CONFIG_NAVIGATOR_FULL_MISSION_CACHE_SIZE
 	MissionLandState _mission_land {};
 	SafePointState _safe_point{};
 };
